created new file just for namespaces
