+++ src/d-d-interfaces.ts
@@ -1,0 +1,14 @@
+// Drag and Drop interfaces
+namespace DDInterfaces {
+    
+}
+interface Draggable {
+    dragStartHandler(event: DragEvent): void;
+    dragEndHandler(event: DragEvent): void;
+}
+
+interface DragTarget {
+    dragOverHandler(event: DragEvent): void;
+    dropHandler(event: DragEvent): void;
+    dragLeaveHandler(event: DragEvent): void;
+}
